requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15.2,<4.0
